Using sk::rt::SystemException in pthreads implementation library. Still does not link because of wrong reloc objects handling.
git-svn-id: https://svn.code.sf.net/p/stemkit/code@1173 faf13d3b-af0d-0410-97ba-fcc57a00092e

+++ lib/sk/rt/thread/pthreads/tests/MutexTest.cc
@@ -12,6 +12,7 @@
 #include "../Mutex.h"
 #include "../Exception.h"
 #include <sk/util/Holder.cxx>
+#include <errno.h>
 
 CPPUNIT_TEST_SUITE_REGISTRATION(sk::rt::thread::pthreads::tests::MutexTest);
 
@@ -48,6 +49,6 @@ testUnlockFailsOnNonLocked()
   }
   catch(const pthreads::Exception& exception) {
     CPPUNIT_ASSERT_EQUAL("pthread_mutex_unlock", exception.getFunctionName());
-    CPPUNIT_ASSERT_EQUAL(EPERM, exception.getCode());
+    CPPUNIT_ASSERT_EQUAL(uint32_t(EPERM), exception.getCode());
   }
 }
